improvement(stress): move out stress code out of longevity test to utils
The 'longevity_test.py' test module has code related to running stress
commands which will be useful in other tests.
So, move it out that logic to the common module 'sdcm/utils/loader_utils.py'.
New mixin class then easily will be able to be added to any test class.
